refactor(orders): extract algo_helpers inline tests to sibling file
Per CLAUDE.md rules 13/14 - tests in their own file, modernize touched
modules. This branch has touched algo_helpers.rs three times adding
helpers; the sibling extraction was overdue. Sibling _tests.rs wired
via #[path] attribute, contents unchanged.
